# functions/src/swagger.yaml
openapi: 3.0.0
info:
  title: USA All Benefits Group Backend API
  description: |
    Backend REST para registro de usuarios, pagos Stripe, cursos, notificaciones WhatsApp,
    recomendaciones, catálogo de servicios y gestión de servicios contratados por clientes.
    Auth propio con JWT. Ejecutado en AWS EC2.
  version: 2.0.0
  contact:
    name: USA All Benefits Group
servers:
  - url: http://ec2-100-48-16-180.compute-1.amazonaws.com:3000
    description: Servidor AWS EC2 (produccion)
  - url: http://localhost:3000
    description: Servidor local (Express)
paths:
...
  /api/auth/register:
    post:
      summary: Registrar un nuevo usuario
      description: |
        Crea una cuenta con email, password, nombre, telefono y rol.
        Roles validos: `client`, `customer`, `student`, `seller`.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  example: "Juan Perez"
                email:
                  type: string
                  format: email
                  example: "juan@example.com"
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: "secreto123"
                phone:
                  type: string
                  nullable: true
                  example: "+13051234567"
                role:
                  type: string
                  enum:
                    - client
                    - customer
                    - student
                    - seller
                  example: "client"
      responses:
        "201":
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  uid:
                    type: string
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      uid:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      phone:
                        type: string
                        nullable: true
                      role:
                        type: string
        "400":
          description: Datos invalidos
        "409":
          description: El email ya esta registrado
  /api/auth/login:
    post:
      summary: Iniciar sesion
      description: Autentica con email y password y devuelve un JWT.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "juan@example.com"
                password:
                  type: string
                  format: password
                  example: "secreto123"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      uid:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      phone:
                        type: string
                        nullable: true
                      role:
                        type: string
        "401":
          description: Credenciales invalidas
  /api/auth/admin/login:
    post:
      summary: Login de administrador
      description: Login con credenciales hardcoded (admin/admin).
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "admin"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      uid:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                      role:
                        type: string
        "401":
          description: Credenciales invalidas
  /api/auth/logout:
    post:
      summary: Cerrar sesion
      description: Revoca el token JWT enviado en el header.
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sesion cerrada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "401":
          description: No autenticado
  /api/auth/me:
    get:
      summary: Obtener perfil del usuario autenticado
      description: Devuelve los datos del usuario segun el token JWT enviado.
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Perfil del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                    nullable: true
                  role:
                    type: string
        "401":
          description: No autenticado
        "404":
          description: Usuario no encontrado
...
  /api/admin/recommendations:
    get:
      summary: Listar todas las recomendaciones (admin)
      description: Devuelve todas las recomendaciones para gestion.
      tags:
        - Recommendations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista completa
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
    post:
      summary: Crear recomendacion (admin)
      description: Crea una nueva recomendacion. Se puede subir imagen como multipart/form-data o base64.
      tags:
        - Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                subtitle:
                  type: string
                type:
                  type: string
                  enum: [instagram, youtube, general]
                externalUrl:
                  type: string
                imageUrl:
                  type: string
                color:
                  type: string
                icon:
                  type: string
                ctaLabel:
                  type: string
                ctaLink:
                  type: string
                active:
                  type: string
                  example: "true"
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Recomendacion creada
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
  /api/admin/recommendations/{id}:
    put:
      summary: Actualizar recomendacion (admin)
      tags:
        - Recommendations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
      responses:
        "200":
          description: Recomendacion actualizada
        "404":
          description: No encontrada
    delete:
      summary: Eliminar recomendacion (admin)
      tags:
        - Recommendations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Eliminada
        "404":
          description: No encontrada
  /api/admin/recommendations/reorder:
    put:
      summary: Reordenar recomendaciones (admin)
      tags:
        - Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      order:
                        type: integer
      responses:
        "200":
          description: Orden actualizado
...
  /api/admin/service-catalog:
    get:
      summary: Listar todo el catalogo (admin)
      description: Incluye activos e inactivos.
      tags:
        - Service Catalog
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Catalogo completo
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
    post:
      summary: Crear item en catalogo (admin)
      description: Agrega un nuevo tipo de servicio/producto al catalogo.
      tags:
        - Service Catalog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "Seguro Dental Individual"
                type:
                  type: string
                  example: "Dental"
                description:
                  type: string
                  example: "Cobertura dental completa"
                active:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Item creado
        "400":
          description: Datos invalidos
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
  /api/admin/service-catalog/{id}:
    put:
      summary: Actualizar item de catalogo (admin)
      tags:
        - Service Catalog
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                description:
                  type: string
                active:
                  type: boolean
      responses:
        "200":
          description: Item actualizado
        "404":
          description: No encontrado
    delete:
      summary: Eliminar item de catalogo (admin)
      tags:
        - Service Catalog
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Eliminado
        "404":
          description: No encontrado
  /api/my-services:
    get:
      summary: Ver mis servicios contratados (cliente)
      description: Devuelve todos los servicios asignados al usuario autenticado. Busca por uid y fallback por email.
      tags:
        - Client Services
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de servicios del cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  services:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        serviceName:
                          type: string
                        serviceType:
                          type: string
                          nullable: true
                        policyNumber:
                          type: string
                          nullable: true
                        contractDate:
                          type: string
                          nullable: true
                        expiryDate:
                          type: string
                          nullable: true
                        status:
                          type: string
                          enum: [active, expired, cancelled, pending]
                        coverageAmount:
                          type: number
                          nullable: true
                        premiumAmount:
                          type: number
                          nullable: true
                        currency:
                          type: string
                          nullable: true
                        notes:
                          type: string
                          nullable: true
                        beneficiaryName:
                          type: string
                          nullable: true
                        beneficiaryPhone:
                          type: string
                          nullable: true
                        createdAt:
                          type: string
                        updatedAt:
                          type: string
        "401":
          description: No autenticado
  /api/admin/services:
    get:
      summary: Listar todos los servicios contratados (admin)
      description: Devuelve todos los servicios de todos los clientes.
      tags:
        - Client Services
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista completa
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
    post:
      summary: Crear servicio para un cliente (admin)
      description: |
        Asigna un servicio individual a un cliente con datos completamente personalizados
        (numero de poliza, fechas, precio, beneficiarios, etc.).
      tags:
        - Client Services
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - serviceName
                - contractDate
              properties:
                userId:
                  type: string
                  description: UID del cliente
                  example: "abc123"
                userEmail:
                  type: string
                  format: email
                  nullable: true
                  example: "cliente@example.com"
                serviceName:
                  type: string
                  description: Nombre del servicio (puede ser del catalogo o personalizado)
                  example: "Indexed Universal Life (IUL)"
                serviceType:
                  type: string
                  nullable: true
                  example: "IUL"
                policyNumber:
                  type: string
                  nullable: true
                  example: "POL-IUL-2024-001"
                contractDate:
                  type: string
                  format: date
                  example: "2024-01-15"
                expiryDate:
                  type: string
                  format: date
                  nullable: true
                  example: "2034-01-15"
                status:
                  type: string
                  enum: [active, expired, cancelled, pending]
                  default: "active"
                coverageAmount:
                  type: number
                  nullable: true
                  example: 500000
                premiumAmount:
                  type: number
                  nullable: true
                  example: 25000
                currency:
                  type: string
                  nullable: true
                  example: "USD"
                notes:
                  type: string
                  nullable: true
                  example: "Cliente legacy importado"
                beneficiaryName:
                  type: string
                  nullable: true
                  example: "Maria Garcia"
                beneficiaryPhone:
                  type: string
                  nullable: true
                  example: "+1234567890"
                catalogItemId:
                  type: string
                  nullable: true
                  description: ID del item del catalogo asociado
      responses:
        "201":
          description: Servicio creado
        "400":
          description: Datos invalidos
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
  /api/admin/services/{id}:
    put:
      summary: Actualizar servicio de cliente (admin)
      description: Edita cualquier campo de un servicio ya asignado.
      tags:
        - Client Services
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Servicio actualizado
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
        "404":
          description: No encontrado
    delete:
      summary: Eliminar servicio de cliente (admin)
      tags:
        - Client Services
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Eliminado
        "401":
          description: No autenticado
        "403":
          description: Sin permisos
        "404":
          description: No encontrado
...
  /api/requestProductInfo:
    post:
      summary: Solicitar informacion de producto
      description: |
        Crea un lead y genera enlace de WhatsApp para contactar al vendedor.
        No requiere autenticacion, pero si se envia un token se asocia al lead.
      tags:
        - Leads
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - productName
                - phoneNumber
                - customerName
              properties:
                productId:
                  type: string
                  example: "prod_001"
                productName:
                  type: string
                  example: "Curso de Seguros"
                phoneNumber:
                  type: string
                  example: "+13051234567"
                customerName:
                  type: string
                  example: "Juan Perez"
                customerEmail:
                  type: string
                  nullable: true
                  example: "juan@example.com"
      responses:
        "200":
          description: Lead creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      ok:
                        type: boolean
                      leadId:
                        type: string
                      message:
                        type: string
        "400":
          description: Datos invalidos
  /api/sendWhatsappNotification:
    post:
      summary: Enviar notificacion por WhatsApp
      description: Envia un mensaje de WhatsApp a un numero especifico.
      tags:
        - Notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - message
              properties:
                phoneNumber:
                  type: string
                  example: "+13051234567"
                message:
                  type: string
                  example: "Hola, tu pago fue confirmado."
      responses:
        "200":
          description: Mensaje enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      ok:
                        type: boolean
                      message:
                        type: string
        "401":
          description: No autenticado
        "400":
          description: Numero o mensaje invalido
  /stripeWebhook:
    post:
      summary: Webhook de Stripe
      description: |
        Recibe eventos de Stripe para confirmar pagos, fallos y reembolsos.
        **No usar desde el frontend.** Stripe llama a este endpoint directamente.
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Evento procesado
        "400":
          description: Firma invalida
        "405":
          description: Metodo no permitido
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /api/auth/login o /api/auth/register